hz from Int to BigInt. good lord.

+++ src/main/scala/ee/hrzn/chryse/platform/Platform.scala
@@ -20,5 +20,5 @@ package ee.hrzn.chryse.platform
 
 trait Platform {
   val id: String
-  val clockHz: Int
+  val clockHz: BigInt
 }
